More complex flusher logic to make default client work
A client now has a default flusher which can be both started
and stopped and the flush interval can be changed dynamically.
This allows for registrering meters before a default flush interval
has been set (which is isn't in the default client).
All meters registred now either has a known flush interval or is assigned
to the default flusher.